requirements.txt
pandas
BeautifulSoup4
requests
linkedin-api
python-dotenv
gspread==3.6